+++ MVCTutorial/Before/Models/Enrollment.cs
@@ -9,9 +9,9 @@ public enum Grade
 
 public class Enrollment
 {
-    public int Id { get; set; }
-    public int CourseId     { get; set; }
-    public int StudentId    { get; set; }
+    public long Id { get; set; }
+    public long CourseId     { get; set; }
+    public long StudentId    { get; set; }
     [DisplayFormat(NullDisplayText = "No grade")]
     public Grade? Grade { get; set; }
